+++ pkg/process.go
@@ -1,4 +1,4 @@
-package process
+package avp
 
 import (
 	"sync"
