+++ scripts/build_protos.sh
@@ -1,3 +1,3 @@
 # !/bin/bash
 
-protoc --go_out=plugins=grpc:. --go_opt=paths=source_relative pkg/proto/avp/avp.proto
+protoc --go_out=plugins=grpc:. --go_opt=paths=source_relative cmd/server/grpc/proto/avp.proto
